firmware: Only default FW_PIC to y if supported
Bare-metal GNU ld does not support PIE, so if using it this this will
result in a failure to build. Instead, default to FW_PIC=n if not
supported. Note that an explicit FW_PIC=y is not overriden, to ensure
the build fails rather than silently producing a position-dependent
binary.

Signed-off-by: Jessica Clarke <jrtc27@jrtc27.com>